pyproject.toml
[tool.poetry]
name = "issue-automation"
version = "0.1.0"
description = ""
authors = ["Deniz Ugur <denizugur@users.noreply.github.com>"]
readme = "README.md"
packages = [{ include = "issue_automation" }]

[tool.poetry.dependencies]
python = "^3.11"
python-docx = "^0.8.11"
pygithub = "^1.59.0"
loguru = "^0.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
process-comments-document = "issue_automation:process_comments_document"
